validated all models

+++ db/migrate/20230816162507_create_reviews.rb
@@ -1,7 +1,7 @@
 class CreateReviews < ActiveRecord::Migration[7.0]
   def change
     create_table :reviews do |t|
-      t.float :rating
+      t.integer :rating
       t.string :content
 
       t.timestamps
